# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Campaign API",
        "description": "API documentation for Campaign management system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/campaigns": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Get a paginated list of campaigns",
                "description": "Retrieve a paginated list of all campaigns. You can control the number of items per page using the per_page parameter.",
                "operationId": "d2a1aecdf4c5fee4d3d5f820b5b2a250",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page (default: 15)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Campaign"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 7
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Create a new campaign",
                "description": "Create a new campaign with the provided data. All required fields must be provided.",
                "operationId": "a0bb4df932d4174de20c996bfa6f3783",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "start_date"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Campaign name",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Summer Sale Campaign"
                                    },
                                    "description": {
                                        "description": "Campaign description",
                                        "type": "string",
                                        "example": "A promotional campaign for summer products",
                                        "nullable": true
                                    },
                                    "start_date": {
                                        "description": "Campaign start date and time",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-06-01T00:00:00Z"
                                    },
                                    "end_date": {
                                        "description": "Campaign end date and time (must be after start_date)",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-08-31T23:59:59Z",
                                        "nullable": true
                                    },
                                    "frequency_minutes": {
                                        "description": "Frequency in minutes (default: 1440 = daily)",
                                        "type": "integer",
                                        "default": 1440,
                                        "minimum": 1,
                                        "example": 1440
                                    },
                                    "status": {
                                        "description": "Campaign status",
                                        "type": "string",
                                        "default": "scheduled",
                                        "enum": [
                                            "scheduled",
                                            "running",
                                            "completed",
                                            "failed"
                                        ],
                                        "example": "scheduled"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Campaign created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Campaign created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Campaign"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The name field is required."
                                                    }
                                                },
                                                "start_date": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The start date field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/campaigns/status/{status}": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Get campaigns by status",
                "description": "Retrieve all campaigns that match the specified status.",
                "operationId": "d6c74ceff7d8fbeea3f2966af1ddc498",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "description": "Campaign status",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "scheduled",
                                "running",
                                "completed",
                                "failed"
                            ],
                            "example": "running"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Campaign"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error - Invalid status value",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "properties": {
                                                "status": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Invalid status value."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Campaign": {
                "description": "Campaign model",
                "properties": {
                    "id": {
                        "description": "Campaign ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Campaign name",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Summer Sale Campaign"
                    },
                    "description": {
                        "description": "Campaign description",
                        "type": "string",
                        "example": "A promotional campaign for summer products",
                        "nullable": true
                    },
                    "start_date": {
                        "description": "Campaign start date and time",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-01T00:00:00Z"
                    },
                    "end_date": {
                        "description": "Campaign end date and time",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-31T23:59:59Z",
                        "nullable": true
                    },
                    "frequency_minutes": {
                        "description": "Frequency in minutes (default: 1440 = daily)",
                        "type": "integer",
                        "example": 1440
                    },
                    "status": {
                        "description": "Campaign status",
                        "type": "string",
                        "enum": [
                            "scheduled",
                            "running",
                            "completed",
                            "failed"
                        ],
                        "example": "scheduled"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}